model/pyproject.toml
[project]
name = "notesagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain>=1.2.15",
    "langchain-core>=1.2.31",
    "langchain-groq>=1.1.2",
    "langgraph>=1.1.6",
    "python-dotenv>=1.2.2",
    "tavily>=1.1.0",
]
